requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-redis==5.4.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy==1.15.0
mypy-extensions==1.0.0
pillow==11.1.0
psycopg==3.2.5
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
segno==1.6.6
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
